feat(control): ADS session state machine
Per-type (LDS/RDS/CDS/EDS) version/nonce tracking, DiscoveryRequest
encoding (initial/ACK/NACK), and response classification through a
consumer-supplied validator — compile failures NACK without advancing
the version, accepts ACK at the new version. Pure logic; the network
driver wires it to h2/gRPC frames. 5 unit tests; vane-control now
depends on vane-proto (leaf, no cycle).

+++ crates/vane-proto/src/pb.rs
@@ -20,7 +20,7 @@ pub enum WireType {
 
 /// Encodes a field tag (field number + wire type).
 pub fn tag(out: &mut Vec<u8>, field: u32, wire: WireType) {
-    varint(out, ((u64::from(field) << 3) | wire as u64) as u64);
+    varint(out, (u64::from(field) << 3) | wire as u64);
 }
 
 /// Appends a base-128 varint.
